Inclusão do Filtro de Órgãos na Gestão da Frota

+++ src/Gesfrota/View/FleetEquipmentForm.php
@@ -17,22 +17,27 @@ use PHPBootstrap\Widget\Form\Controls\TextBox;
 use PHPBootstrap\Widget\Nav\NavLink;
 use PHPBootstrap\Widget\Nav\TabPane;
 use PHPBootstrap\Widget\Nav\Tabbable;
+use PHPBootstrap\Widget\Modal\Modal;
+use PHPBootstrap\Widget\Misc\Title;
+use PHPBootstrap\Widget\Button\Button;
+use PHPBootstrap\Widget\Modal\TgModalClose;
+use PHPBootstrap\Widget\Form\Controls\Decorator\Seek;
+use PHPBootstrap\Widget\Form\Controls\SearchBox;
+use Gesfrota\Model\Domain\Agency;
 
 class FleetEquipmentForm extends AbstractForm {
     
 	
 	/**
 	 * @param Action $submit
-	 * @param Action $seekVehiclePlate
-	 * @param Action $seekVehicleModel
-	 * @param Action $searchVehicleModel
-	 * @param Action $seekOwner
-	 * @param Action $searchOwner
 	 * @param Action $cancel
+	 * @param Action $seekAgency
+	 * @param Action $searchAgency
+	 * @param boolean $showAgencies
 	 * @param ServiceCardForm $subform
 	 */
-    public function __construct(Action $submit, Action $cancel, ServiceCardForm $subform = null ) {
-	    $this->buildPanel('Minha Frota', 'Gerenciar Equipamento');
+	public function __construct(Action $submit, Action $cancel, Action $seekAgency, Action $searchAgency, $showAgencies = false, ServiceCardForm $subform = null ) {
+	    $this->buildPanel('Minha Frota', 'Gerenciar Veículos e Equipamentos');
 		$form = $this->buildForm('fleet-equipment-form');
 		
 		$general = new Fieldset('Identificação do Equipamento');
@@ -47,6 +52,25 @@ class FleetEquipmentForm extends AbstractForm {
 		$input->setSpan(7);
 		$input->setRequired(new Required(null, 'Por favor, preencha esse campo'));
 		$form->buildField('Descrição', $input, null, $general);
+		
+		if ($showAgencies) {
+			$modal = new Modal('agency-search', new Title('Órgãos', 3));
+			$modal->setWidth(600);
+			$modal->addButton(new Button('Cancelar', new TgModalClose()));
+			$form->append($modal);
+			
+			$input = [];
+			$input[0] = new TextBox('agency-id');
+			$input[0]->setSuggestion(new Seek($seekAgency));
+			$input[0]->setRequired(new Required(null, 'Por favor, preencha esse campo'));
+			$input[0]->setSpan(1);
+			
+			$input[1] = new SearchBox('agency-name', $searchAgency, $modal);
+			$input[1]->setEnableQuery(false);
+			$input[1]->setSpan(6);
+			
+			$form->buildField('Órgão', $input, null, $general);
+		}
 		
 		$input = new ComboBox('engine');
 		$input->setSpan(2);
@@ -91,6 +115,10 @@ class FleetEquipmentForm extends AbstractForm {
 	public function extract( Equipment $object ) {
 	    $data['asset-code'] = $object->getAssetCode();
 	    $data['description'] = $object->getDescription();
+	    if ($object->getResponsibleUnit()) {
+	    	$data['agency-id'] = $object->getResponsibleUnit()->getCode();
+	    	$data['agency-name'] = $object->getResponsibleUnit()->getName();
+	    }
 		$data['engine'] = $object->getEngine();
 		$data['fleet'] = $object->getFleet();
 		$data['active'] = $object->getActive();
@@ -99,6 +127,7 @@ class FleetEquipmentForm extends AbstractForm {
 			$data['created-at'] = $object->getCreatedAt()->format('d/m/Y H:m:s');
 			$data['updated-at'] = $object->getUpdatedAt()->format('d/m/Y H:m:s');
 		}
+		
 		$this->component->setData($data);
 	}
 
@@ -112,23 +141,27 @@ class FleetEquipmentForm extends AbstractForm {
 		$object->setEngine((int) $data['engine']);
 		$object->setFleet((int) $data['fleet']);
 		$object->setActive($data['active']);
-		
-		$oldcards = $object->getAllCards();
-		foreach( $data['cards'] as $dto ) {
-			if ($dto->getId()) {
-				unset($oldcards[$dto->getId()]);
-				$card = $em->find(ServiceCard::getClass(), $dto->getId());
-			} else {
-				$card = new ServiceCard();
-				$em->persist($card);
-			}
-			$card->setNumber($dto->getNumber());
-			$provider = $em->find(ServiceProvider::getClass(), $dto->getServiceProvider()->getId());
-			$card->setServiceProvider($provider);
-			$object->addCard($card);
+		if (isset($data['agency-id'])) {
+			$object->setResponsibleUnit($em->find(Agency::getClass(), $data['agency-id']));
 		}
-		foreach ( $oldcards as $card ) {
-			$em->remove($card);
+		if (isset($data['cards'])) {
+			$oldcards = $object->getAllCards();
+			foreach( $data['cards'] as $dto ) {
+				if ($dto->getId()) {
+					unset($oldcards[$dto->getId()]);
+					$card = $em->find(ServiceCard::getClass(), $dto->getId());
+				} else {
+					$card = new ServiceCard();
+					$em->persist($card);
+				}
+				$card->setNumber($dto->getNumber());
+				$provider = $em->find(ServiceProvider::getClass(), $dto->getServiceProvider()->getId());
+				$card->setServiceProvider($provider);
+				$object->addCard($card);
+			}
+			foreach ( $oldcards as $card ) {
+				$em->remove($card);
+			}
 		}
 	}
